refactor: extract ChatFooter into own component and file
